Backend API/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
click==8.1.7
psycopg2-binary
fastapi
uvicorn
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
click==8.1.7
psycopg2-binary
pyasn1==0.6.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
pyasn1==0.6.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
